% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advice_fn.R
\name{advice_fn}
\alias{advice_fn}
\title{Generate catch advice}
\usage{
advice_fn(em, pro.yr = assess.interval)
}
\arguments{
\item{em}{estimation model}

\item{pro.yr}{Number of years in projection}
}
\value{
a list of catch adive
}
\description{
a function to generate catch advice for management strategy evaluation.
}
\seealso{
\code{\link{project_wham}}
}
