% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wham.vis.R
\name{assay}
\alias{assay}
\title{SummarizedExperiment objects}
\usage{
assay(...)
}
\arguments{
\item{...}{arguments in package SummarizedExperiment. See \code{\link[SummarizedExperiment]{assay}}}
}
\description{
SummarizedExperiment objects
}
