% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convergence.R
\name{check_convergence}
\alias{check_convergence}
\title{Check convergence of WHAM model}
\usage{
check_convergence(mod, ret = FALSE)
}
\arguments{
\item{mod}{output from \code{\link{fit_wham}}}

\item{ret}{T/F, return list? Default = FALSE, just prints to console}
}
\value{
a list with three components:
  \describe{
    \item{\code{$convergence}}{From \code{\link[stats:nlminb]{stats::nlminb}}, "0 indicates successful convergence"}
    \item{\code{$maxgr}}{Max gradient, from `max(mod$gr(mod$opt$par))`}
    \item{\code{$maxgr_par}}{Name of parameter with max gradient}
  }
}
\description{
Access quick convergence checks from `TMB` and `nlminb`.
}
\examples{
\dontrun{
data("SNEMA_ytl") # load SNEMA yellowtail flounder data and parameter settings
mod = fit_wham(input) # using default values
check_convergence(mod)
}
}
\seealso{
\code{\link{fit_wham}}, \code{\link{fit_tmb}}, \code{\link[stats:nlminb]{stats::nlminb}}
}
