% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tmb.R
\name{check_estimability}
\alias{check_estimability}
\title{Check for identifiability of fixed effects
Originally provided by https://github.com/kaskr/TMB_contrib_R/TMBhelper 
Internal function called by \code{\link{fit_tmb}}.}
\usage{
check_estimability(obj, h)
}
\arguments{
\item{obj}{The compiled object}

\item{h}{optional argument containing pre-computed Hessian matrix}
}
\value{
A tagged list of the hessian and the message
}
\description{
\code{check_estimability} calculates the matrix of second-derivatives of the marginal likelihood
w.r.t. fixed effects, to see if any linear combinations are not estimable (i.e. cannot be
uniquely estimated conditional upon model structure and available data, e.g., resulting
in a likelihood ridge and singular, non-invertable Hessian matrix)
}
