% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_wham_models.R
\name{compare_wham_models}
\alias{compare_wham_models}
\title{Compare WHAM models}
\usage{
compare_wham_models(
  mods,
  fname = "model_comparison",
  sort = TRUE,
  calc.rho = TRUE,
  calc.aic = TRUE,
  fdir = getwd(),
  do.print = TRUE
)
}
\arguments{
\item{mods}{(named) list of fit WHAM models. If no names are given, "m1", "m2", ...
will be used.}

\item{fname}{character, filename to save CSV results table (.csv will be appended). Default = "model_comparison".}

\item{sort}{T/F, sort by AIC? Default = TRUE.}

\item{calc.rho}{T/F, calculate Mohn's rho? Retrospective analysis must have been run for all modes. Default = TRUE.}

\item{calc.aic}{T/F, calculate AIC? Default = TRUE.}

\item{fdir}{character, path to directory to save CSV results table. Default = getwd().}

\item{do.print}{T/F, print table to console? Default = TRUE.}
}
\value{
a list with the following components:
  \describe{
    \item{\code{daic}}{Vector of delta-AIC by model}
    \item{\code{aic}}{Vector of AIC by model}
    \item{\code{rho}}{Matrix of Mohn's rho by model}
    \item{\code{best}}{Name of best model (lowest AIC)}
    \item{\code{tab}}{Results table of AIC and Mohn's rho}
  }
}
\description{
After fitting multiple WHAM models, \code{compare_wham_models} produces a table
of AIC and Mohn's rho to aid model comparison.
}
\examples{
\dontrun{
m1 <- fit_wham(input1)
m2 <- fit_wham(input2)
mods <- list(m1=m1, m2=m2)
res <- compare_wham_models(mods)
res$best
res$tab
}

}
\seealso{
\code{\link{fit_wham}}
}
