% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_retro_peels.R
\name{do_retro_peels}
\alias{do_retro_peels}
\title{Fit retrospective peels and add them to the fitted model object}
\usage{
do_retro_peels(
  model,
  n.peels = 7,
  ran = NULL,
  use.mle = TRUE,
  do.sdrep = FALSE,
  n.newton = 0,
  MakeADFun.silent = FALSE,
  retro.silent = FALSE,
  save.input = FALSE,
  do.brps = FALSE,
  check.version = TRUE
)
}
\arguments{
\item{model}{Optimized TMB model, output from \code{\link{fit_tmb}}.}

\item{n.peels}{Integer, number of peels to use in retrospective analysis. Default = \code{7}.}

\item{ran}{Character, specifies which parameters to treat as random effects. Default = \code{"model$input$random"}.}

\item{use.mle}{T/F, use MLEs from full model fit as initial values for each peel? If not, the initial values from full model input are used. Default = \code{TRUE}.}

\item{do.sdrep}{T/F, calculate standard deviations of model parameters for each peel? Default = \code{FALSE}.}

\item{n.newton}{integer, number of additional Newton steps after optimization for each peel. Default = \code{0}.}

\item{MakeADFun.silent}{T/F, Passed to silent argument of \code{\link[TMB:MakeADFun]{TMB::MakeADFun}}. Default = \code{FALSE}.}

\item{retro.silent}{T/F, Passed to argument of internal fit_peel function. Determines whether peel number is printed to screen. Default = \code{FALSE}.}

\item{save.input}{T/F, should modified input list be saved for every peel? Necessary to project from a peel but increases model object size. Default = \code{FALSE}.}

\item{do.brps}{T/F, calculate and report biological reference points}

\item{check.version}{T/F, whether to verify the wham package commit and version for the fitted model are the same as the currently used package.}
}
\value{
\code{peels}, a list of length \code{n.peels}, where entry \emph{i} is a model
fit by peeling off \emph{i} years of data.
}
\description{
Function to add fitted retrospective peels to fitted model returned by \code{\link{fit_wham}}. Calls \code{\link{retro}}.
to fit the model peeling off \code{1, 2, ..., n.peels} years of data.
This is just a wrapper for retro that instead of returning just the list of peels, returns the fitted model with the peels.
}
\seealso{
\code{\link{fit_wham}}, \code{\link{retro}}, \code{\link{fit_peel}}
}
