% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tmb.R
\name{extract_fixed}
\alias{extract_fixed}
\title{Extract fixed effects
Originally provided by https://github.com/kaskr/TMB_contrib_R/TMBhelper 
Internal function called by \code{\link{check_estimability}}.}
\usage{
extract_fixed(obj)
}
\arguments{
\item{obj, }{The compiled object}
}
\value{
A vector of fixed-effect estimates
}
\description{
\code{extract_fixed} extracts the best previous value of fixed effects, in a way that works for both mixed and fixed effect models
}
