% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.wham.fn.R
\name{fit.wham.fn}
\alias{fit.wham.fn}
\title{Fit WHAM model}
\usage{
fit.wham.fn(input, n.newton = 3, do.sdrep = TRUE, do.retro = TRUE,
  n.peels = 7)
}
\arguments{
\item{input}{Named list with several components:
\describe{
  \item{\code{input$data}}{Data to fit the assessment model to.}
  \item{\code{input$par}}{Parameters, a list of all parameter objects required by the user template (both random and fixed effects). See \code{\link[TMB]{MakeADFun}}.}
  \item{\code{input$map}}{Map, a mechanism for collecting and fixing parameters. See \code{\link[TMB]{MakeADFun}}.}
  \item{\code{input$random}}{Character vector defining the random effect parameters. See \code{\link[TMB]{MakeADFun}}.}
  \item{\code{input$years}}{Numeric vector of the years to run the assessment for}
  \item{\code{input$map.n}}{I don't know what this is.}
  \item{\code{input$model_name}}{Character, name of the model, e.g. \code{"Yellowtail flounder"}}
  \item{\code{input$ages.lab}}{Character vector of the age labels, e.g. \code{c("1","2","3","4+").}}
}}

\item{n.newton}{integer, number of additional Newton steps after optimization. Passed to \code{\link{fit.tmb.fn}}. Default = \code{3}.}

\item{do.sdrep}{T/F, calculate standard deviations of model parameters? See \code{\link[TMB]{sdreport}}. Default = \code{TRUE}.}

\item{do.retro}{T/F, do retrospective analysis? Default = \code{TRUE}.}

\item{n.peels}{integer, number of peels to use in retrospective analysis. Default = \code{7}.}
}
\value{
\code{mod}, a fit TMB model
}
\description{
Fits the compiled WHAM model using \code{\link[TMB:MakeADFun]{TMB::MakeADFun}} and
\code{\link[stats:nlminb]{stats::nlminb}}. Runs retrospective analysis if specified.
}
\examples{
\dontrun{
data("SNEMA_ytl") # load SNEMA yellowtail flounder data and parameter settings
mod = fit.wham.fn(input) # using default values
}
}
\seealso{
\code{\link{fit.tmb.fn}}, \code{\link{retro.fn}}
}
