% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tmb.R
\name{fit_tmb}
\alias{fit_tmb}
\title{Fit TMB model using nlminb}
\usage{
fit_tmb(model, n.newton = 3, do.sdrep = TRUE, do.check = FALSE)
}
\arguments{
\item{model}{Output from \code{\link[TMB:MakeADFun]{TMB::MakeADFun}}.}

\item{n.newton}{Integer, number of additional Newton steps after optimization. Default = \code{3}.}

\item{do.sdrep}{T/F, calculate standard deviations of model parameters? See \code{\link[TMB]{sdreport}}. Default = \code{TRUE}.}

\item{do.check}{T/F, check if model parameters are identifiable? Runs \code{\link[TMBhelper::Check_Identifiable]{TMBhelper::Check_Identifiable}}. Default = \code{TRUE}.}
}
\value{
\code{model}, appends the following:
  \describe{
    \item{\code{model$opt}}{Output from \code{\link[stats:nlminb]{stats::nlminb}}}
    \item{\code{model$date}}{System date}
    \item{\code{model$dir}}{Current working directory}
    \item{\code{model$rep}}{model$report()}
    \item{\code{model$TMB_version}}{Version of TMB installed}
    \item{\code{model$parList}}{List of parameters, \code{model$env$parList()}}
    \item{\code{model$final_gradient}}{Final gradient, \code{model$gr()}}
    \item{\code{model$sdrep}}{Estimated standard deviations for model parameters, \code{\link[TMB:sdreport]{TMB::sdreport}}}
  }
}
\description{
Runs optimization on the TMB model using \code{\link[stats:nlminb]{stats::nlminb}}.
If specified, takes additional Newton steps and calculates standard deviations.
Internal function called by \code{\link{fit_wham}}.
}
\seealso{
\code{\link{fit_wham}}, \code{\link{retro}}, \code{\link{TMBhelper::Check_Identifiable}}
}
