% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_move.R
\name{fix_move}
\alias{fix_move}
\title{Fix movement parameter}
\usage{
fix_move(input)
}
\arguments{
\item{input}{wham input}
}
\value{
a wham input with movement rate treated as known
}
\description{
a function to fix movement rate in the input.
}
