% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_F_from_Catch_region.R
\name{get_F_from_Catch_region}
\alias{get_F_from_Catch_region}
\title{Calculate region-specific F from region-specific catch}
\usage{
get_F_from_Catch_region(
  Catch,
  NAA,
  log_M,
  mu,
  L,
  sel,
  fracyr_season,
  fleet_regions,
  fleet_seasons,
  can_move,
  mig_type,
  waacatch,
  trace,
  F_init
)
}
\arguments{
\item{Catch}{Number of regions}

\item{NAA}{n_stocks x n_regions x ny x n_ages; numbers-at-age}

\item{log_M}{n_stocks x n_regions x ny x n_ages; natural mortality}

\item{mu}{n_stocks x n_ages x n_seasons x n_years x n_regions x n_regions; movement rates}

\item{L}{L: n_years x n_regions; "extra" mortality rate}

\item{sel}{Model years (or "burn-in" period in MSE)}

\item{fracyr_season}{n_seasons; length of intervals for each season}

\item{fleet_regions}{n_fleets; which region each fleet is operating}

\item{fleet_seasons}{n_fleets x n_seasons; 0/1 indicating whether fleet is operating in the season}

\item{can_move}{n_stocks x n_seasons x n_regions x n_regions; 0/1 determining whether movement can occur from one region to another}

\item{mig_type}{n_stocks; 0 = migration after survival, 1 = movement and mortality simultaneous}

\item{waacatch}{n_fleets x n_ages; weight-at-age}

\item{trace}{trace: 0/1 whether to print stuff to screen}

\item{F_init}{n_regions}
}
\value{
Region-specific F estimated using newton method and nlminb method
}
\description{
The get_F_from_Catch_region() function is used to calculate region-specific F given the aggregate region-specific catch
Note that some helper functions that support the get_F_from_Catch_region() function are included below.
}
