% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_through_fn.R
\name{loop_through_fn}
\alias{loop_through_fn}
\title{Function to perform management strategy evaluation}
\usage{
loop_through_fn(
  om,
  M_om,
  sel_om,
  NAA_re_om,
  mean_rec_weights = NULL,
  move_om,
  age_comp_om = "multinomial",
  M_em,
  sel_em,
  NAA_re_em,
  move_em,
  em.opt = NULL,
  age_comp_em = "multinomial",
  assess_years = NULL,
  assess_interval = NULL,
  base_years = NULL,
  year.use = 30,
  hcr.type = 1,
  hcr.opts = NULL,
  do.retro = FALSE,
  do.osa = FALSE,
  seed = 123,
  save.sdrep = FALSE
)
}
\arguments{
\item{om}{Operating model including pseudo data}

\item{M_om}{Natural mortality configuration in the operating model}

\item{sel_om}{Selectivity configuration in the operating model}

\item{NAA_re_om}{Numbers-at-age (NAA) configuration in the operating model}

\item{mean_rec_weights}{A vector of weights given to each stock (default = NULL).}

\item{move_om}{Movement configuration in the operating model}

\item{age_comp_om}{Likelihood distribution of age composition data in the operating model
\itemize{
  \item \code{"multinomial"} Default
  \item \code{"dir-mult"}
  \item \code{"dirichlet-miss0"}
  \item \code{"dirichlet-pool0"}
  \item \code{"dir-mult"}
  \item \code{"logistic-normal-miss0"}    
  \item \code{"logistic-normal-ar1-miss0"}
  \item \code{"logistic-normal-pool0"}
  \item \code{"logistic-normal-01-infl"}
  \item \code{"logistic-normal-pool0"}
  \item \code{"logistic-normal-01-infl-2par"}
  \item \code{"mvtweedie"}
  \item \code{"dir-mult-linear"}
  }}

\item{M_em}{Natural mortality configuration in the assessment model}

\item{sel_em}{Selectivity configuration in the assessment model}

\item{NAA_re_em}{Numbers-at-age (NAA) configuration in the assessment model}

\item{move_em}{Movement configuration in the assessment model}

\item{em.opt}{Type of assessment model
\itemize{
  \item \code{$separate.em} TRUE = spatially implicit, FALSE = spatially disaggregated
  \item \code{$separate.em.type} when separate.em = TRUE \cr
  {=1} separate/independent assessment models (i.e. global SPR brps = FALSE) \cr
  {=2} fleets-as-areas (global SPR brps = TRUE) \cr
  {=3} panmictic (use spatially-aggregated data and projected total catch will be allocated based on stock-specific recruitment) \cr
  \item \code{$do.move} T/F movement is included (use when separate.em = FALSE)
  \item \code{$est.move} T/F movement rate is estimated (use when separate.em = FALSE)
  }}

\item{age_comp_em}{Likelihood distribution of age composition data in the assessment model
\itemize{
  \item \code{"multinomial"} (Default)
  \item \code{"dir-mult"}
  \item \code{"dirichlet-miss0"}
  \item \code{"dirichlet-pool0"}
  \item \code{"dir-mult"}
  \item \code{"logistic-normal-miss0"}    
  \item \code{"logistic-normal-ar1-miss0"}
  \item \code{"logistic-normal-pool0"}
  \item \code{"logistic-normal-01-infl"}
  \item \code{"logistic-normal-pool0"}
  \item \code{"logistic-normal-01-infl-2par"}
  \item \code{"mvtweedie"}
  \item \code{"dir-mult-linear"}
  }}

\item{assess_years}{Year in which the assessment is conducted}

\item{assess_interval}{Assessment interval used in the MSE feedback loop}

\item{base_years}{Years used in the burn-in period}

\item{year.use}{Number of years used in the assessment model}

\item{hcr.type}{Type of harvest control rule
\itemize{
  \item \code{"1"} Annual projected catch based on 75% of F40% (default)
  \item \code{"2"} Constant catch based on 75% of F40%
  \item \code{"3"} "Hockey stick" catch based on stock status
  }}

\item{hcr.opts}{a list of HCR information, only used if hcr.type = 3
\itemize{
  \item \code{"max_percent"} maximum percent of F_XSPR to use for calculating catch in projections (default = 75)
  \item \code{"min_percent"} minimum percent of F_XSPR to use for calculating catch in projections (default = 0.01)
  \item \code{"BThresh_up"} Upper bound of overfished level (default = 0.5) 
  \item \code{"BThresh_low"} Lower bound of overfished level (default = 0.1)
  }}

\item{do.retro}{T/F do retrospective analysis? Default = TRUE.}

\item{do.osa}{T/F calculate one-step-ahead (OSA) residuals? Default = TRUE.}

\item{seed}{Seed used for generating data}

\item{save.sdrep}{T/F save the full report of the assessment model (memory intensive)}
}
\value{
a list of model output
  \describe{
    \item{\code{$om}}{Operating model}
    \item{\code{$sdrep_list}}{Parameter estimates from each assessment model}
    \item{\code{$em_list}}{Short report of each assessment model}
    \item{\code{$opt_list}}{Output from \code{\link[stats:nlminb]{stats::nlminb}}}
    \item{\code{$converge_list}}{Convergence from each assessment model}
    \item{\code{$em_full}}{Full report of each assessment model}
  }
}
\description{
a wrapper function to step through the feedback period and update the operating model, 
and refit the estimation model and generate the catch advice.
}
\seealso{
\code{\link{make_em_input}}, \code{\link{update_om_fn}}, \code{\link{advice_fn}}
}
