% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_through_fn2.R
\name{loop_through_fn2}
\alias{loop_through_fn2}
\title{Function to perform management strategy evaluation}
\usage{
loop_through_fn2(
  om,
  M_em,
  sel_em,
  NAA_re_em,
  move_em = NULL,
  em.opt = NULL,
  assess_years = NULL,
  assess_interval = NULL,
  base_years = NULL,
  year.use = 30,
  seed = 123,
  save.sdrep = FALSE
)
}
\arguments{
\item{om}{Operating model}

\item{M_em}{Natural mortality random effects}

\item{sel_em}{Selectivity random effects}

\item{NAA_re_em}{Numbers-at-age random effects}

\item{move_em}{Movement random effects}

\item{em.opt}{Movement random effects
\itemize{
  \item \code{$separate.em} TRUE = spatially implicit, FALSE = spatially disaggregated
  \item \code{$separate.em.type} when separate.em = TRUE \cr
  {=1} fleets-as-areas (global SPR brps = FALSE) \cr
  {=2} fleets-as-areas (global SPR brps = TRUE) \cr
  {=3} panmictic (spatially-aggregated) \cr
  \item \code{$do.move} T/F movement is included (use when separate.em = FALSE)
  \item \code{$est.move} T/F movement rate is estimated (use when separate.em = FALSE)
  }}

\item{assess_years}{Year in which the assessment is conducted}

\item{assess_interval}{Assessment interval used in the MSE feedback loop}

\item{base_years}{Years used in the burn-in period}

\item{year.use}{Number of years used in the assessment model}

\item{seed}{Seed used for generate data}

\item{save.sdrep}{T/F save the full report (memory intensive)}
}
\value{
a list of model output
  \describe{
    \item{\code{$om}}{Operating model}
    \item{\code{$sdrep_list}}{Parameter estimates from each assessment model}
    \item{\code{$em_list}}{Short report of each assessment model}
    \item{\code{$opt_list}}{Output from \code{\link[stats:nlminb]{stats::nlminb}}}
    \item{\code{$converge_list}}{Convergence from each assessment model}
    \item{\code{$om_full}}{Full report of each assessment model}
  }
}
\description{
a wrapper function to step through the feedback period and update the operating model, 
and refit the estimation model and generate the catch advice.
}
\seealso{
\code{\link{make_em_input}}, \code{\link{update_om_fn}}, \code{\link{advice_fn}}
}
