% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_osa_residuals.R
\name{make_osa_residuals}
\alias{make_osa_residuals}
\title{Calculate one-step-ahead residuals}
\usage{
make_osa_residuals(
  model,
  osa.opts = list(method = "oneStepGaussianOffMode", parallel = TRUE)
)
}
\arguments{
\item{model}{A fit WHAM model, output from \code{\link{fit_wham}}.}
}
\value{
the same fit TMB model with additional elements for osa residuals:
  \describe{
    \item{\code{$OSA.Ecov}}{data.frame returned by \code{\link[TMB]{TMB::oneStepPredict}} for environmental observations, if applicable.}
    \item{\code{$OSA.agregate}}{data.frame returned by \code{\link[TMB]{TMB::oneStepPredict}} for aggregate catch and index observations, if applicable.}
    \item{\code{$OSA.agecomp}}{data.frame returned by \code{\link[TMB]{TMB::oneStepPredict}} for age composition observations, if applicable.}
    \item{\code{$osa}}{One-step-ahead residuals (if \code{do.osa=TRUE})}
  }
}
\description{
Standard residuals are not appropriate for models with random effects. Instead, one-step-ahead (OSA) residuals
can be used for evaluating model goodness-of-fit (\href{https://link.springer.com/article/10.1007/s10651-017-0372-4}{Thygeson et al. (2017)},
implemented in \code{\link[TMB:oneStepPredict]{TMB::oneStepPredict}}). Certain OSA residual options
are passed to \code{\link[TMB:oneStepPredict]{TMB::oneStepPredict}} in a list \code{osa.opts}. See details below.
It is not recommended to run this function (or \code{\link[TMB:oneStepPredict]{TMB::oneStepPredict}}) with any random effects and
mvtweedie age composition likelihoods as the extensive computational demand. An error will be thrown in such cases.
}
\examples{
\dontrun{
data("input4_SNEMAYT") # load SNEMA yellowtail flounder data and parameter settings
mod <- fit_wham(input4_SNEMAYT, do.osa =FALSE, do.retro =FALSE)
mod <- make_osa_residuals(mod) # calculate Mohn's rho
plot_wham_output(mod)
}

}
\seealso{
\code{\link{fit_wham}}
}
