% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peel.fit.fn.R
\name{peel.fit.fn}
\alias{peel.fit.fn}
\title{Fit model peeling off \emph{i} years of data}
\usage{
peel.fit.fn(peel, model, do.sdrep = FALSE, n.newton = 3)
}
\arguments{
\item{peel}{Integer, number of years of data to remove before model fitting.}

\item{model}{Output from \code{\link{fit.tmb.fn}}.}

\item{do.sdrep}{T/F, calculate standard deviations of model parameters? Default = \code{FALSE}.}

\item{n.newton}{integer, number of additional Newton steps after optimization for each peel. Default = \code{3}.}
}
\value{
\code{out}, output of \code{\link{fit.tmb.fn}} for peel \emph{i}
}
\description{
Internal function called by \code{\link{retro.fn}} for \emph{i} in 1--\code{n.peels}.
Fits the model peeling off \emph{i} years of data (calls \code{\link{fit.tmb.fn}}).
}
\seealso{
\code{\link{fit.wham.fn}}, \code{\link{retro.fn}}, \code{\link{fit.tmb.fn}}
}
