% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mse_output.R
\name{plot_mse_output}
\alias{plot_mse_output}
\title{Plot performance metrics for one realization using different management strategies}
\usage{
plot_mse_output(mods, main.dir = getwd(), use.n.years = 10, dpi = 150)
}
\arguments{
\item{mods}{A list of models}

\item{main.dir}{Path to save the final report}

\item{use.n.years}{Number of years that is used to summarize statistics}

\item{dpi}{Resolution (number of dots per inches)}
}
\value{
a report
}
\description{
a function to compare performance metrics of different assessment models
}
