% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_wham.R
\name{project_wham}
\alias{project_wham}
\title{Project WHAM model}
\usage{
project_wham(mod, n.years = 3)
}
\arguments{
\item{mod}{list, output from \code{\link{fit_wham}}}

\item{n.years}{integer, number of years to project (default = 3)}
}
\value{
the input list \code{mod}, with an additional element \code{proj}.
\code{mod$proj} is a fit TMB model, same as output by \code{\link{fit_wham}}, except
with additional elements \code{mod$proj$proj_IAA} (projected index of abundance at-age, mean)
and \code{mod$proj$proj_IAA$sd} (projected index of abundance at-age, sd).
}
\description{
After fitting a WHAM model, \code{project_wham} projects the model forward.
}
\examples{
\dontrun{
m1 <- fit_wham(input1)
m1 <- project_wham(m1, n.years = 3)
}

}
\seealso{
\code{\link{fit_wham}}
}
