% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wham_fit.R
\name{read_wham_fit}
\alias{read_wham_fit}
\title{Read WHAM fit}
\usage{
read_wham_fit(mod, alphaCI = 0.05)
}
\arguments{
\item{mod}{output from \code{\link{fit_wham}}}

\item{alphaCI}{(1-alpha)\% confidence intervals will be calculated. Default = 0.05 for 95\% CI.}
}
\value{
a named list with the following elements:
  \describe{
    \item{\code{$years}}{numeric vector, model years only, e.g. \code{1972:2020}}
    \item{\code{$years_full}}{numeric vector, model + proj years, e.g. \code{1972:2022}}
    \item{\code{$selAA}}{list of length(n_selblocks), first the fleet blocks then indices, i.e. if 4 fleet blocks and 3 indices, \code{selAA[[5]]} is for index 1. Each element is a matrix, years (rows) x ages (cols), selectivity at age}
    \item{\code{$selblock_pointer_fleets}}{matrix, years x fleets, indices of selAA used by each fleet in each year}
    \item{\code{$selblock_pointer_indices}}{matrix, years x indices, indices of selAA used by each index in each year}
    \item{\code{$MAA}}{array, stocks x regions x years x ages, natural mortality}
    \item{\code{$log_SSB}}{matrix, years x 2, log-scale spawning stock biomass. 1st col = MLE, 2nd col = SE.}
    \item{\code{$log_F}}{matrix, years x 2, log-scale fully-selected F. 1st col = MLE, 2nd col = SE.}
    \item{\code{$log_NAA_rep}}{array, stocks x regions x years x ages, numbers at age}
    \item{\code{$NAA_CV}}{array, stocks x regions x years x ages, CV of numbers at age}
    \item{\code{$percentSPR}}{scalar, X\% SPR used to calculate reference points, default = 40}
    \item{\code{$log_Y_FXSPR}}{matrix, years x 2, log-scale yield at FXSPR. 1st col = MLE, 2nd col = SE.}
    \item{\code{$log_FXSPR}}{matrix, years x 2, log-scale FXSPR. 1st col = MLE, 2nd col = SE.}
    \item{\code{$log_SSB_FXSPR}}{matrix, years x 2, log-scale SSB at FXSPR. 1st col = MLE, 2nd col = SE.}
    \item{\code{$log_rel_ssb_F_cov}}{list, length n_years, each element is a 2x2 covariance matrix with SSB/SSB_FXSPR first and F/F_FXSPR second}
  }
}
\description{
Gets output from a fit WHAM model for plotting with other models.
Internal function, called within \code{\link{compare_wham_models}}.
}
\seealso{
\code{\link{fit_wham}}, \code{\link{read_asap3_fit}}, \code{\link{compare_wham_models}}
}
