% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro.res.fn.R
\name{retro.res.fn}
\alias{retro.res.fn}
\title{Extract retrospective results for plotting}
\usage{
retro.res.fn(model)
}
\arguments{
\item{model}{A fit WHAM model, output from \code{\link{fit.wham.fn}} with \code{do.retro = TRUE}.}
}
\value{
a named list with the components:
  \describe{
    \item{\code{SSB}}{Spawning stock biomass}
    \item{\code{Fbar}}{Fishing mortality}
    \item{\code{NAA}}{Numbers-at-age}
 }
}
\description{
Extract retrospective results for plotting
}
\examples{
\dontrun{
data("SNEMA_ytl") # load SNEMA yellowtail flounder data and parameter settings
mod = fit.wham.fn(input) # using default values: do.retro = T, n.peels = 7
x = retro.res.fn(mod) # get retrospective results
}
}
\seealso{
\code{\link{fit.wham.fn}}, \code{\link{retro.fn}}
}
