% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro_res.R
\name{retro_res}
\alias{retro_res}
\title{Extract retrospective results for plotting}
\usage{
retro_res(model)
}
\arguments{
\item{model}{A fit WHAM model, output from \code{\link{fit_wham}} with \code{do.retro = TRUE}.}
}
\value{
a named list with the components:
  \describe{
    \item{\code{SSB}}{Spawning stock biomass}
    \item{\code{Fbar}}{Fishing mortality}
    \item{\code{NAA}}{Numbers-at-age}
 }
}
\description{
Extract retrospective results for plotting
}
\examples{
\dontrun{
data("SNEMA_ytl") # load SNEMA yellowtail flounder data and parameter settings
mod = fit_wham(input) # using default values: do.retro = T, n.peels = 7
x = retro_res(mod) # get retrospective results
}
}
\seealso{
\code{\link{fit_wham}}, \code{\link{retro}}
}
