% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.fn.R
\name{rho.fn}
\alias{rho.fn}
\title{Calculate Mohn's rho for a WHAM model with peels}
\usage{
rho.fn(model)
}
\arguments{
\item{model}{A fit WHAM model, output from \code{\link{fit.wham.fn}} with \code{do.retro = TRUE}.}
}
\value{
\code{rho}, a vector of Mohn's rho
}
\description{
Calculate Mohn's rho for a WHAM model with peels
}
\examples{
\dontrun{
data("SNEMA_ytl") # load SNEMA yellowtail flounder data and parameter settings
mod = fit.wham.fn(input) # using default values: do.retro = T, n.peels = 7
rho.fn(mod) # calculate rho
}
}
\seealso{
\code{\link{fit.wham.fn}}, \code{\link{retro.fn}}
}
