% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_age_sel0.R
\name{set_age_sel0}
\alias{set_age_sel0}
\title{Make one or more selectivity blocks with age-specific parameters}
\usage{
set_age_sel0(input, selblocks)
}
\arguments{
\item{input}{list containing data and parameters (output from \code{\link{prepare_wham_input}})}

\item{selblocks}{numeric, number of age-specific selectivity blocks}
}
\value{
a modified list of data and parameters
}
\description{
Make one or more selectivity blocks with age-specific parameters
}
\examples{
\dontrun{
asap3 = read_asap3_dat("ASAP_SNEMAYT.dat")
input = prepare_wham_input(asap3)
input = set_age_sel0(input, selblocks = 1:3)
mod = fit_wham(input)
}

}
