% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ecov.R
\name{set_ecov}
\alias{set_ecov}
\title{Specify configuration for environmental covariates, effects on the population, and parameter values}
\usage{
set_ecov(input, ecov)
}
\arguments{
\item{input}{list containing data, parameters, map, and random elements (output from \code{\link{wham::prepare_wham_input}})}

\item{ecov}{(optional) named list of environmental covariate data and parameters (see details)

\code{ecov} specifies any environmental covariate data and model as well as the effect on the population. Environmental covariate data need not span
the same years as the fisheries data. It can be \code{NULL} if no environmental data are to be fit.
Otherwise, it must be a named list with the following components:
  \describe{
    \item{$label}{Name(s) of the environmental covariate(s). Used in printing.}
    \item{$mean}{Mean observations (matrix). number of years x number of covariates. Missing values = NA.}
    \item{$logsigma}{Configure observation standard errors. Options:
      \describe{
        \item{Matrix of \eqn{log} standard errors with same dimensions as \code{$mean}}{Specified values for each time step }
        \item{log standard errors for each covariate, numeric vector or matrix w/ dim 1 x n.ecov}{Specified value the same for all time steps}
        \item{estimation option (for all covariates). character string:}{
          \code{"est_1"}: Estimated, one value shared among time steps.
          \code{"est_re"}: Estimated value for each time step as random effects with two parameters (mean, var)}
        \item{list of two elements.}{
          First is the matrix of log standard errors or the vector of single values for each covariate as above. 
          Second is a character vector of estimation options (\code{NA}, \code{"est_1"},\code{"est_re"}) for each covariate. 
          For covariates with non-NA values, values in the first element are ignored.}
      }
    }
    \item{$year}{Years corresponding to observations (vector of same length as \code{$mean} and \code{$logsigma})}
    \item{$use_obs}{T/F (or 1/0) vector/matrix of the same dimension as \code{$mean} and \code{$logsigma}.
    Use the observation? Can be used to ignore subsets of the ecov without changing data files.}
    \item{$process_model}{Process model for the ecov time-series. \code{"rw"} = random walk, \code{"ar1"} = 1st order autoregressive, 
       \code{NA} = do not fit}
    \item{$process_mean_vals}{vector of (initial) mean values for the ecov time-series.}
    \item{$process_sig_vals}{vector of (initial) standard deviation values for the ecov time-series.}
    \item{$process_cor_vals}{vector of (initial) correlation values for the ecov time-series.}
    \item{$recruitment_how}{character matrix (n_Ecov x n_stocks) indicating how each ecov affects recruitment for each stock. 
       Options are based on (see \href{https://www.sciencedirect.com/science/article/pii/S1385110197000221}{Iles & Beverton (1998)}) 
       combined with the order of orthogonal polynomial of the covariate and has the form "type-lag-order". "type" can be:
        \describe{
          \item{= "none"}{no effect.}
          \item{= "controlling"}{pre-recruit density-independent mortality.}
          \item{= "limiting"}{ maximum recruitment, e.g. ecov determines amount of suitable habitat)}
          \item{= "lethal"}{threshold, i.e. R --> 0 at some ecov value.}
          \item{= "masking"}{metabolic/growth, decreases dR/dS}
          \item{= "directive"}{e.g. behavioral}
         }
        for type other than "none", "lag" can be:
        \describe{
          \item{= "lag-n"}{lag = n which can be 0,1,2,.... lag-1 implies the covariate in year y affects recruitment in year y+1.}
        }
        for "type" being other than "none", "order" can be:
        \describe{
          \item{= "linear"}{the covariate effect is linear on the transformed recruitment parameter (e.g., log).}
          \item{= "poly-n"}{orthogonal polynomial where n = 1 (same as "linear"),2,...}
        }
        so "limiting-lag-1-poly-2" would model the covariate affecting recruitment the next year (lag = 1) as a second order orthogonal 
         polynomial (\eqn{b_0 + b_1*ecov + b_2*ecov^2 + ...}) limiting effect.
     }
    \item{$M_how}{character array (n_Ecov x n_stocks x n_ages x n_regions) indicating how each ecov affects M by age,stock,region and 
       has the form "lag-order". "lag" can be:
        \describe{
          \item{= "none"}{no effect.}
          \item{= "lag-n"}{lag = n which can be 0,1,2,.... lag-1 implies the covariate in year y affects M in year y+1.}
        }
        for "lag" being other than "none", "order" can be:
        \describe{
          \item{= "linear"}{the covariate effect is linear on the transformed M parameter (e.g., log).}
          \item{= "poly-n"}{orthogonal polynomial where n = 1 (same as "linear"),2,...}
         }
      }
    \item{$M_effect_map}{integer array (n_stocks x n_ages x n_regions x n_Ecov) indicating which estimated effects are common by age,stock,region.
      If not specified there the same effect is estimated for all M where $M_how is other than "none" for each covariate.}
    \item{$q_how}{character matrix (n_Ecov x n_indices) indicating whether each ecov affects catchability for each index. and has 
     the form "lag-order". "lag" can be:
        \describe{
          \item{= "none"}{no effect.}
          \item{= "lag-n"}{lag = n which can be 0,1,2,.... lag-1 implies the covariate in year y affects catchability in year y+1.}
        }
        for "lag" being other than "none", "order" can be:
        \describe{
          \item{= "linear"}{the covariate effect is linear on the transformed catchability parameter (e.g., log).}
          \item{= "poly-n"}{orthogonal polynomial where n = 1 (same as "linear"),2,...}
         }
      }
    \item{$move_how}{character array (n_Ecov x n_stocks x n_ages x n_seasons x n_regions x n_regions - 1) indicating whether each ecov 
       affects movement from one region to the others by stock,age,season. and has the form "lag-order". "lag" can be:
        \describe{
          \item{= "none"}{no effect.}
          \item{= "lag-n"}{lag = n which can be 0,1,2,.... lag-1 implies the covariate in year y affects a movement parameter in year y+1.}
        }
        for "lag" being other than "none", "order" can be:
        \describe{
          \item{= "linear"}{the covariate effect is linear on the transformed movement parameter (e.g., log).}
          \item{= "poly-n"}{orthogonal polynomial where n = 1 (same as "linear"),2,...}
         }
       }
    \item{$move_effect_map}{integer array (n_stocks x n_ages x n_seasons x n_regions x n_regions-1 x n_Ecov) indicating which estimated 
       effects are common by age,stock,region, season etc. If not specified the same effect is estimated for all movement parameters
       where $move_how is other than "none" for each covariate.}
    \item{$beta_R_vals}{n_stocks x n_ecov x max(n_poly_R) array of initial values for effects on recruitment.}
    \item{$beta_M_vals}{n_stocks x n_ages x n_regions x n_ecov x max(n_poly_M) array of initial values for effects on natural mortality.}
    \item{$beta_q_vals}{n_indices x n_ecov x max(n_poly_q) array of initial values for effects on catchability.}
    \item{$beta_mu_vals}{n_stocks x n_ages x n_seasons x n_regions x n_regions - 1 x n_ecov x max(n_poly_move) array of initial values for effects on movement parameters.}
  }}
}
\value{
a named list with same elements as the input provided with environmental covariate observations, effects, and model options modified.
}
\description{
Specify configuration for environmental covariates, effects on the population, and parameter values
}
\examples{
\dontrun{
wham.dir <- find.package("wham")
path_to_examples <- system.file("extdata", package="wham")
asap3 <- read_asap3_dat(file.path(path_to_examples,"ex1_SNEMAYT.dat"))
env.dat <- read.csv(file.path(path_to_examples,"GSI.csv"), header=T)
input <- prepare_wham_input(asap3, NAA_re = list(sigma = "rec"))
ecov <- list(
 label = "GSI",
 mean = as.matrix(env.dat$GSI),
 logsigma = 'est_1', # estimate obs sigma, 1 value shared across years
 year = env.dat$year,
 use_obs = matrix(1, ncol=1, nrow=dim(env.dat)[1]), # use all obs (=1)
 process_model = 'ar1') # "rw" or "ar1"
input <- set_ecov(input, ecov = ecov) #GSI in the model without any effects
}

}
\seealso{
\code{\link{prepare_wham_input}}
}
