% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_F.R
\name{set_F}
\alias{set_F}
\title{Specify configuration for fully-selected fishing mortality}
\usage{
set_F(input, F_opts = NULL)
}
\arguments{
\item{input}{list containing data, parameters, map, and random elements (output from \code{\link{wham::prepare_wham_input}})}

\item{F_opts}{(optional) named list of initial values for annual fully-selected fishing mortality and configuration method for estimation.

\code{F_opts} specifies a few as well as the effect on the population. Environmental covariate data need not span
the same years as the fisheries data. It can be \code{NULL} if no environmental data are to be fit.
Otherwise, it must be a named list with the following components:
  \describe{
    \item{$F}{matrix (n_years x n_fleets) of (initial) values for fully-selected fishing morality.}
    \item{$F_config}{integer 1: (default) configure F parameters (on log scale) as an F in the initial year and then deviations from one year to the next,
       or 2: configure F parameters as (log) annual values.}
 }}
}
\description{
Specify configuration for fully-selected fishing mortality
}
