% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_indices.R
\name{set_indices}
\alias{set_indices}
\title{Specify index selectivity blocks and aggregate and age composition observations for indices}
\usage{
set_indices(input, index_info = NULL)
}
\arguments{
\item{input}{list containing data, parameters, map, and random elements (output from \code{\link{prepare_wham_input}})}

\item{index_info}{(optional) list specifying various aspects about catch by indices (see details)

\code{index_info} specifies observations, and various configuration options for index-specific catch observations and will overwrite attributes specified in the ASAP data file.
If \code{NULL}, all settings from the ASAP data file or basic_info are used.
\code{index_info} is a list with any of the following entries:
  \describe{
    \item{$n_indices}{number of indices}
    \item{$index_regions}{vector (n_indices) of regions where each fleet operates.}
    \item{$index_seasons}{vector (n_indices) of 0/1 values flagging which seasons each index occurs.}
    \item{$agg_indices}{matrix (n_years_model x n_indices) of annual aggregate index catches.}
    \item{$agg_index_cv}{matrix (n_years_model x n_indices) of CVs for annual aggregate index catches.}
    \item{$fracyr_indices}{matrix (n_years_model x n_indices) of fractions of year at which index occurs within the season (difference between time of survey and time at start of season).}
    \item{$use_indices}{matrix (n_years_model x n_indices) of 0/1 values flagging whether to use aggregate observations.}
    \item{$units_indices}{matrix (n_years_model x n_indices) of 1/2 values flagging whether aggregate observations are biomass (1) or numbers (2).}
    \item{$index_paa}{array (n_indices x n_years_model x n_ages) of annual catch proportions at age by index.}
    \item{$use_index_paa}{matrix (n_years_model x n_indices) of 0/1 values flagging whether to use proportions at age observations.}
    \item{$units_index_paa}{matrix (n_years_model x n_indices) of 1/2 values flagging whether composition observations are biomass (1) or numbers (2).}
    \item{$index_Neff}{matrix (n_years_model x n_indices) of effective sample sizes for proportions at age observations.}
    \item{$waa_pointer_indices}{vector (n_indices) of itegers indicated waa to use for each index.}
    \item{$selblock_pointer_indices}{matrix (n_years_model x n_indices) of itegers indicated selblocks to use.}
    \item{$initial_index_sd_scale}{vector (n_indices) of scalar multipliers of annual log-observation standard deviation. Default = 1.}
    \item{$map_index_sd_scale}{integer vector (n_indices) specifying which sd scalar parameters to fix. Use \code{NA} to fix a parameter and integers to estimate. 
        Use the same integer for multiple indices to estimate a shared scalar parameter.}
  }}
}
\description{
Specify index selectivity blocks and aggregate and age composition observations for indices
}
