% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_L.R
\name{set_L}
\alias{set_L}
\title{Specify model and parameter configuration for "extra" mortality not directly attributed to natural mortality}
\usage{
set_L(input, L)
}
\arguments{
\item{input}{list containing data, parameters, map, and random elements (output from \code{\link{wham::prepare_wham_input}})}

\item{L}{(optional) list specifying "extra" mortality options: model, random effects, initial values, and parameters to fix (see details)

\code{L} specifies estimation options for "extra" mortality.
If \code{NULL}, This mortality source is not used. 
\code{L} is a list with the following entries:
  \describe{
    \item{$model}{length = n_regions. "extra" mortality model options are:
                   \describe{
                     \item{"none"}{(default) no extra mortality for this region.}
                     \item{"constant"}{estimate a single mean mortality for the region shared across all ages}
                     \item{"iid_re"}{estimate independent random effects over years, for the region}
                     \item{"ar1_re"}{estimate random effect correlated over years, for the region}
                   }
                 }
    \item{$initial_means}{Initial/mean L-at-region}
    \item{$sigma_vals}{Initial standard deviation by region value to use for the L random effects. Values are not used if \code{L$model} = "none".}
    \item{$cor_vals}{Initial correlation values to use for the L random effects. If unspecified all initial values are 0}
  }}
}
\description{
Specify model and parameter configuration for "extra" mortality not directly attributed to natural mortality
}
