% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_move.R
\name{set_move}
\alias{set_move}
\title{Specify model and parameter configuration for movement when input$data$n_regions > 1}
\usage{
set_move(input, move)
}
\arguments{
\item{input}{list containing data, parameters, map, and random elements (output from \code{\link{wham::prepare_wham_input}})}

\item{move}{(optional) list specifying movement options: model, random effects, initial values, and parameters to fix (see details)

\code{move} specifies estimation options for movement.
If \code{NULL}, no movement will occur. If there are multiple regions, each stock will be modeled separately in different regions without movement. 
\code{move} is a list with the following entries:
  \describe{
    \item{$stock_move}{length = n_stocks, T/F whether each stock can move. If not provided then movement will be defined below for all stocks.}
    \item{$separable}{length = n_stocks, T/F whether movement should be modeled separably from mortality or both occuring simultaneously.}
    \item{$mean_model}{matrix (n_regions x (n_regions-1)): model options for fixed effects (mean and possibly variance) for each movement parameter are:
                   \describe{
                     \item{"none"}{(default) no movement between regions.}
                     \item{"constant"}{estimate a single movement rate to each region shared across all stocks, seasons, ages, years}
                     \item{"season"}{estimate movement rates  to each region for each season shared across all stocks, ages, years}
                     \item{"stock_constant"}{estimate a movement rate for each stock to each region shared across all seasons, ages, years}
                     \item{"stock_season"}{estimate a movement rate for each stock each season to each region shared across all ages, years}
                   }
    }
    \item{$age_re}{matrix (n_regions x (n_regions-1)): options for age random effects (for each mean parameter defined in \code{move$mean_model}):
                   \describe{
                     \item{"none"}{(default) no movement rate random effects by age.}
                     \item{"iid"}{independent movement rate random effects by age.}
                     \item{"ar1"}{allow first order autoregressive correlation of movement rate random effects by age.}
                   }
    }
    \item{$year_re}{matrix (n_regions x (n_regions-1)): options for yearly random effects (for each mean parameter defined in \code{move$mean_model}):
                   \describe{
                     \item{"none"}{(default) no movement rate random effects by year.}
                     \item{"iid"}{independent movement rate random effects by year.}
                     \item{"ar1"}{allow first order autoregressive correlation of movement rate random effects by year.}
                   }
    }
    \item{$prior_sigma}{array (n_stocks x n_seasons x n_regions x n_regions - 1) of sd parameters for normal priors on mean movement parameters on transformed scale (-Inf,Inf)}
    \item{$use_prior}{array (n_stocks x n_seasons x n_regions x n_regions - 1) 0/1 indicator whether to include prior for mean movement parameters in joint log-likelihood.}
    \item{$can_move}{array (n_stocks x n_seasons x n_regions x n_regions) 0/1 indicator whether movement can occur from one region to another.}
    \item{$must_move}{array (n_stocks x n_seasons x n_regions) 0/1 indicator whether movement from region must occur.}
    \item{$mean_vals}{array (n_stocks x n_seasons x n_regions x n_regions-1) of initial movement rate parameters *from* each region. Usage depends on \code{move$mean_model}.}
    \item{$sigma_vals}{array (n_stocks x n_seasons x n_regions x n_regions -1) of initial standard deviations to use for random effects. Usage depends on \code{move$age_re} and \code{move$year_re}.}
    \item{$cor_vals}{array (n_stocks x n_seasons x n_regions x n_regions - 1x 2) of initial correlation values to use for random effects. Usage depends on \code{move$age_re} and \code{move$year_re}.
       cor_vals[,,,,1] is for correlation with age, and cor_vals[,,,,2] is for correlation with year.}
  }}
}
\description{
Specify model and parameter configuration for movement when input$data$n_regions > 1
}
