% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_osa_obs.R
\name{set_osa_obs}
\alias{set_osa_obs}
\title{Set up observation vector that is used by the model for likelihood calculations and one-step-ahead residuals.}
\usage{
set_osa_obs(input)
}
\arguments{
\item{input}{list containing data, parameters, map, and random elements (output from \code{\link{wham::prepare_wham_input}})}
}
\value{
the same input list as provided, but with $obs and $obsvec configured.
This is run after any changes have been made to the data
}
\description{
Set up observation vector that is used by the model for likelihood calculations and one-step-ahead residuals.
}
