% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wham.vis.R
\name{TaxaHeatmap}
\alias{TaxaHeatmap}
\title{TaxaHeatmap}
\usage{
TaxaHeatmap(
  object,
  taxa.level = c("k", "p", "o", "c", "f", "g", "s", "otu"),
  filter = c("all", "custom", "DE.filter"),
  custom,
  p.cutoff = 0.05,
  effect.size.range = c(0, 0),
  scale = T,
  col,
  column_split = F,
  annotation,
  count = c("rawCount", "WhamTransformation"),
  ...
)
}
\arguments{
\item{object}{required. The output of WhamBiobakery, WhamEBI or Wham16s}

\item{taxa.level}{Required, c("k","p","o","c","f","g","s").Collapse taxa at the provided level.}

\item{filter}{c("all","custom","DE.filter"). Default is "all". "all" shows all samples. "custom" shows taxa provided by user and all samples. "DE.filter" shows taxa which pass DE test cutoff(p.cutoff and effect.size.range) and samples which are in the DE.result contrast.}

\item{custom}{required when filter is set as "custom". A vector which contain taxa to be shown}

\item{p.cutoff}{applied when filter is set as "DE.filter". Default is 0.05.}

\item{effect.size.range}{applied when filter is set as "DE.filter" and aldex.module is set as t.test.Default range is c(0,0)}

\item{scale}{default is TRUE. Z score is used when set as T. Expression counts are used when set as F}

\item{col}{an argument in ComplexHeatmap::Heatmap: vector of colors if the color mapping is discrete or a color mapping function if the matrix is continuous numbers (should be generated by colorRamp2). If the matrix is continuous, the value can also be a vector of colors so that colors can be interpolated.}

\item{column_split}{Split on columns.TRUE, FALSE or a vector provided by users.Default value is False. Column will be spilt based on annotation.}

\item{annotation}{optional. Default is the last variable in design. A character vector indicates what column of metadata will be used for annotation bar}

\item{count}{c("rawCount","WhamTransformation"). Default value is "rawCount". Setting as "rawCount" uses original count. Setting as WhamTransformation converts count to ratio(count divided by the total sum of each sample), and then times one million( x 1000,000)}

\item{...}{arguments in ComplexHeatmap::Heatmap}
}
\value{
a heatmap of taxonomy
}
\description{
Visualization of Wham upstream analysis
}
\examples{
# see the vignette for more details
}
