% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wham.support.R
\name{TaxaIden_fn}
\alias{TaxaIden_fn}
\title{Workflow Hub for Automated Metagenomic Exploration (WHAM!)}
\usage{
TaxaIden_fn(TaxaInfo, taxa.level = c("k", "p", "o", "c", "f", "g", "s", "otu"))
}
\arguments{
\item{TaxaInfo}{a vector store taxonomic information}

\item{taxa.level}{required, c("k","p","o","c","f","g","s"). Collapsing the count to provided taxonomic level. options are "k","p","c","o","f","g","s".Only need to declaimed when argument DE is setup as "taxa". Every taxonimic level has to have "k__","p__","c__","o__","f__","g__",or "s__"}
}
\value{
a character vector corresponds with taxa.level
}
\description{
check the format of taxainfo and output taxa.LEVEL
}
\examples{
taxainfo = c("Unclassified",
             "k__Archaea;p__Euryarchaeota;c__Methanobacteria;o__Methanobacteriales",
             "k__Archaea;p__Euryarchaeota;c__Methanobacteria;o__Methanobacteriales;f__Methanobacteriaceae")
TaxaIden_fn(taxainfo,taxa.level = "f")

}
