% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wham.support.R
\name{TaxaToDF_fn}
\alias{TaxaToDF_fn}
\title{Workflow Hub for Automated Metagenomic Exploration (WHAM!)}
\usage{
TaxaToDF_fn(taxainfo, level_table = c("not_otu", "otu"))
}
\arguments{
\item{taxainfo}{a vector store taxonomic information}

\item{level_table}{decide to use 16s or shotgun sequencing}
}
\value{
return a data frame contain taxonomic information
}
\description{
converting taxa character to data frame. Missing taxa are filled with 'unassiagned'
}
\examples{
taxainfo = c("Unclassified",
             "k__Archaea;p__Euryarchaeota;c__Methanobacteria;o__Methanobacteriales",
             "k__Archaea;p__Euryarchaeota;c__Methanobacteria;o__Methanobacteriales;f__Methanobacteriaceae")
TaxaToDF_fn(taxainfo)

}
