% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_wham_input.R
\name{update_wham_input}
\alias{update_wham_input}
\title{Update input data and parameters from simulated data with projection}
\usage{
update_wham_input(simres, model, n_years_add)
}
\arguments{
\item{simres}{list containing data and parameters (output from model$simulate(complete=TRUE), where model is output from \code{fit_wham})}

\item{model}{object used to produce simres.}

\item{n_years_add}{number of projection years to convert to model_years}
}
\value{
a named list with the following components:
  \describe{
    \item{\code{data}}{Named list of data, passed to \code{\link[TMB:MakeADFun]{TMB::MakeADFun}}}
    \item{\code{par}}{Named list of parameters, passed to \code{\link[TMB:MakeADFun]{TMB::MakeADFun}}}
    \item{\code{map}}{Named list defining how to optionally collect and fix parameters, passed to \code{\link[TMB:MakeADFun]{TMB::MakeADFun}}}
    \item{\code{random}}{Character vector of parameters to treat as random effects, passed to \code{\link[TMB:MakeADFun]{TMB::MakeADFun}}}
    \item{\code{years}}{Numeric vector of years to fit WHAM model}
    \item{\code{years_full}}{Numeric vector of years to fit and project WHAM model}
    \item{\code{ages.lab}}{Character vector of age labels, ending with plus-group}
    \item{\code{model_name}}{Character, name of stock/model (specified in call to \code{prepare_wham_input})}
  }
}
\description{
When data are simulated from a wham model with projections, this function can be used to update the input model years.
This is helpful for using WHAM as an assessment model in a management strategy evaluation. See \code{prepare_wham_input}, 
\code{prepare_wham_om_input}, \code{prepare_wham_om_proj}, and \code{project_wham} for further details on data and parameter structure.
}
\examples{
\dontrun{
input = prepare_wham_om_input(asap3)
mod = fit_wham(input, do.fit = FALSE, do.proj = TRUE)
sim = mod$simulate(complete=TRUE)
sim_input = update_wham_input(sim, mod, 3)
simfit = fit_wham(sim_input, MakeADFun.silent = TRUE)
plot_wham_output(simfit)
}

}
\seealso{
\code{\link{read_asap3_dat}}, \code{\link{fit_wham}}, \href{https://www.nefsc.noaa.gov/nft/ASAP.html}{ASAP}, \href{https://www.sciencedirect.com/science/article/pii/S1385110197000221}{Iles & Beverton (1998)}
}
