% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wham.vis.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{VolcanoPlot}
\usage{
VolcanoPlot(
  object = NULL,
  effect = c("effect"),
  p_value = c("we.eBH", "we.ep", "wi.eBH", "wi.ep"),
  pCutoff = 0.05,
  effectCutoff = 1,
  ...
)
}
\arguments{
\item{object}{required. The output of WhamBiobakery, WhamEBI or Wham16s}

\item{effect}{c("effect"), Use effect as effectcutoff. Default is "effect"}

\item{p_value}{c("we.eBH","we.ep","wi.eBH","wi.ep"). Using one of them as p-value.Default is "we.eBH".}

\item{pCutoff}{Cut-off for statistical significance. A horizontal line will be drawn at -log10(pCutoff). DEFAULT = 0.05. OPTIONAL}

\item{effectCutoff}{ut-off for absolute effect size. Vertical lines will be drawn at the negative and positive values of FCCutoff. DEFAULT = 1.0. OPTIONAL.}

\item{...}{...}
}
\value{
a volcano plot of statistical result
}
\description{
Visualization of Wham upstream analysis
}
\examples{
# see the vignette for more details
}
