% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wham_html.R
\name{wham_html}
\alias{wham_html}
\title{Create HTML file to view output plots in browser}
\usage{
wham_html(dir.main = NULL, title = "WHAM Output", width = 500, openfile = TRUE)
}
\arguments{
\item{dir.main}{directory to save html file (\code{\link{plot_wham_output}}
makes `.png` plot files in a `plots_png` subdirectory of `dir.main`).}

\item{title}{Title for HTML page.}

\item{width}{Width of plots (in pixels).}

\item{openfile}{Automatically open index.html in default browser?}
}
\description{
Writes a set of HTML files with tabbed navigation between them. Called by
\code{\link{plot_wham_output}} if `out.type = 'html'` (default). Opens main file in
default browser. Modified from [`r4ss::SS_html`](https://github.com/r4ss/r4ss/blob/master/R/SS_html.R).
}
\seealso{
\code{\link{plot_wham_output}}, `r4ss::SS_html()`
}
