% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wham.core.R
\name{WhamDEFromEBI}
\alias{WhamDEFromEBI}
\title{workflow Hub for Automated Metagenomic Exploration (WHAM!)}
\usage{
WhamDEFromEBI(
  object = NULL,
  DE = c("feature", "taxa"),
  contrast,
  ref.contrast,
  All.contrast,
  taxa.level = c("k", "p", "o", "c", "f", "g", "s"),
  WhamTransformation = c("count", "WhamTransformation"),
  useMC = F,
  aldex.module = c("t.test", "anova"),
  anova.group
)
}
\arguments{
\item{object}{required. Object generated from WhamFromBiobakery, WhamFromEBI or WhamFrom16s.}

\item{DE}{required. Type of differential expression. Options are "taxa",or "feature".}

\item{contrast}{Specifying what comparison to extract from the WhamInput. a vector indicates the variable, numerator,and denominator. e.g. contrast = c("group","eye","mouth"). The argument contrast, ref.contrast, and All.contrast can't be declaimed at the same time. If none of three are declaimed, function will use last variable in design as the variable, the first element in the variable as the numerator, and the last element in the variable as the denominator.}

\item{ref.contrast}{Specifying a reference element in the variable, All other elements in the variable will be processed one vs one comparison (vs reference variable). A vector indicates the variable, denominator. e.g. contrast = c("cytokine","Buffer"). The argument 'contrast', 'ref.contrast', and 'All.contrast' can't be declaimed at the same time. If none of three are declaimed, function will use last element in design as the variable, the first element in the variable as the numerator, and the last element in the variable as the denominator.}

\item{All.contrast}{Specifying a variable. Every two elements in the variable will be processed one vs one comparison. a vector indicates the variable, denominator. e.g. contrast = c("cytokine"). The argument 'contrast', 'ref.contrast', and 'All.contrast' can't be declaimed at the same time. If none of three are declaimed, function will use last element in design as the variable, the first element in the variable as the numerator, and the last element in the variable as the denominator.}

\item{taxa.level}{required, c("k","p","o","c","f","g","s"). Collapsing the count to provided taxonomic level. options are "k","p","c","o","f","g","s".Only need to declaimed when argument DE is setup as "taxa". Every taxonimic level has to have "k__","p__","c__","o__","f__","g__",or "s__"}

\item{WhamTransformation}{c("count","WhamTransformation"). Default value is "count". Setting as "count" uses original count. Setting as WhamTransformation converts count to ratio(count divided by the total sum of each sample), and then times one million( x 1000,000)}

\item{useMC}{use multicore by default(FALSE). Multi core processing will be attempted with the BiocParallel package, then the parallel package. If neither are installed, serial processing will be used.}

\item{aldex.module}{c("t.test","anova"), Default value is t.test(Welch's t and Wilcoxon rank test). "anova" will perform glm and Kruskal Wallace tests for one-way ANOVA}

\item{anova.group}{Specifying what variable in design to be performed anova analysis. if anova.group is NULL, the last variable in design will be used.}
}
\value{
a dataframe contains statistic results or a WhamResult object contains statistic results and contrast.

we.ep - Expected P value of Welch's t test

we.eBH - Expected Benjamini-Hochberg corrected P value of Welch's t test

wi.ep - Expected P value of Wilcoxon rank test

wi.eBH - Expected Benjamini-Hochberg corrected P value of Wilcoxon test

rab.all: a vector containing the median clr value for each feature

rab.win.conditionA: a vector containing the median clr value for each feature in condition A

rab.win.conditionB: a vector containing the median clr value for each feature in condition B

diff.btw: a vector containing the per-feature median difference between condition A and B

diff.win: a vector containing the per-feature maximum median difference between Dirichlet instances within conditions

effect:  median effect size: diff.btw / max(diff.win) for all instances

overlap: proportion of effect size that overlaps 0 (i.e. no effect)

kw.ep a vector containing the expected p-value of the Kruskal-Wallis test for each feature

kw.eBH a vector containing the corresponding expected value of the Benjamini-Hochberg corrected p-value for each feature

glm.ep a vector containing the expected p-value of the glm ANOVA for each feature

glm.eBH a vector containing the corresponding expected value of the Benjamini-Hochberg corrected p-value for each feature
}
\description{
R package allowing for exploratory analysis of metagenomics and metatranscriptomic data.  Includes visualization and statistical analysis on the gene family and taxa level. This package is developped based on package ALDEx2.
}
\examples{
# see the vignette for more details
}
