% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wham.core.R
\name{WhamEBI}
\alias{WhamEBI}
\title{workflow Hub for Automated Metagenomic Exploration (WHAM!)}
\usage{
WhamEBI(
  featureCount = NULL,
  taxaCount = NULL,
  colData = NULL,
  design = NULL,
  DE = c("taxa", "feature"),
  taxa.level = c("k", "p", "o", "c", "f", "g", "s"),
  contrast,
  ref.contrast,
  All.contrast,
  WhamTransformation = c("count", "WhamTransformation"),
  useMC = F,
  aldex.module = c("t.test", "anova"),
  anova.group
)
}
\arguments{
\item{featureCount}{a data frame or a matrix contains the feature counts generated from the EBI process. The first two columns of the count table have to be "Acc" ,and "Feature" respectively(case insensitive)}

\item{taxaCount}{a data frame or a matrix contains the taxa counts generated from the EBI process. The first column of the count table have to be "taxa" respectively(case insensitive)}

\item{colData}{metadata table contains experiment design with at least one column. The row number of coldata has to be equal to the columns of count table excluding the first three columns.}

\item{design}{a formula expresses how the counts for each genes depend on the variables.e.g ~ group}

\item{DE}{required. Type of differential expression. Options are "taxa",or "feature".}

\item{taxa.level}{required, c("k","p","o","c","f","g","s"). Collapsing the count to provided taxonomic level. options are "k","p","c","o","f","g","s".Only need to declaimed when argument DE is setup as "taxa". Every taxonimic level has to have "k__","p__","c__","o__","f__","g__",or "s__"}

\item{contrast}{Specifying what comparison to extract from the WhamInput. a vector indicates the variable, numerator,and denominator. e.g. contrast = c("group","eye","mouth"). The argument contrast, ref.contrast, and All.contrast can't be declaimed at the same time. If none of three are declaimed, function will use last variable in design as the variable, the first element in the variable as the numerator, and the last element in the variable as the denominator.}

\item{ref.contrast}{Specifying a reference element in the variable, All other elements in the variable will be processed one vs one comparison (vs reference variable). A vector indicates the variable, denominator. e.g. contrast = c("cytokine","Buffer"). The argument 'contrast', 'ref.contrast', and 'All.contrast' can't be declaimed at the same time. If none of three are declaimed, function will use last element in design as the variable, the first element in the variable as the numerator, and the last element in the variable as the denominator.}

\item{All.contrast}{Specifying a variable. Every two elements in the variable will be processed one vs one comparison. a vector indicates the variable, denominator. e.g. contrast = c("cytokine"). The argument 'contrast', 'ref.contrast', and 'All.contrast' can't be declaimed at the same time. If none of three are declaimed, function will use last element in design as the variable, the first element in the variable as the numerator, and the last element in the variable as the denominator.}

\item{WhamTransformation}{c("count","WhamTransformation"). Default value is "count". Setting as "count" uses original count. Setting as WhamTransformation converts count to ratio(count divided by the total sum of each sample), and then times one million( x 1000,000)}

\item{useMC}{use multicore by default(FALSE). Multi core processing will be attempted with the BiocParallel package, then the parallel package. If neither are installed, serial processing will be used.}

\item{aldex.module}{c("t.test","anova"), Default value is t.test(Welch's t and Wilcoxon rank test). "anova" will perform glm and Kruskal Wallace tests for one-way ANOVA}

\item{anova.group}{Specifying what variable in design to be performed anova analysis. if anova.group is NULL, the last variable in design will be used.}
}
\value{
a dataframe contains statistic results or a WhamResult object contains statistic results and contrast.

we.ep - Expected P value of Welch's t test

we.eBH - Expected Benjamini-Hochberg corrected P value of Welch's t test

wi.ep - Expected P value of Wilcoxon rank test

wi.eBH - Expected Benjamini-Hochberg corrected P value of Wilcoxon test

rab.all: a vector containing the median clr value for each feature

rab.win.conditionA: a vector containing the median clr value for each feature in condition A

rab.win.conditionB: a vector containing the median clr value for each feature in condition B

diff.btw: a vector containing the per-feature median difference between condition A and B

diff.win: a vector containing the per-feature maximum median difference between Dirichlet instances within conditions

effect:  median effect size: diff.btw / max(diff.win) for all instances

overlap: proportion of effect size that overlaps 0 (i.e. no effect)

kw.ep a vector containing the expected p-value of the Kruskal-Wallis test for each feature

kw.eBH a vector containing the corresponding expected value of the Benjamini-Hochberg corrected p-value for each feature

glm.ep a vector containing the expected p-value of the glm ANOVA for each feature

glm.eBH a vector containing the corresponding expected value of the Benjamini-Hochberg corrected p-value for each feature
}
\description{
workflow Hub for Automated Metagenomic Exploration (WHAM!)
}
\examples{
dir.feature <-  system.file("extdata","EBI_feature_input.tsv",package = "Wham")
feature_EBI <-  read.delim(dir.feature)[,1:32]
dir.taxa <-  system.file("extdata","EBI_taxa_input.tsv",package = "Wham")
taxa_EBI <-  read.delim(dir.taxa)[,1:31]

##create fake metadata information
metadata = data.frame(group = rep(c("arm","mouth","eye"),c(10,10,10)),
                      age = sample(15:50,30,
                      replace = TRUE))
rownames(metadata) = colnames(feature_EBI)[3:32]

WhamEBI(featureCount = feature_EBI,
       taxaCount = taxa_EBI,
       colData = metadata,
       design = ~ group,
       DE ="taxa",
       taxa.level = "g")

# see the vignette for more details

}
