% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wham.core.R
\name{WhamFrom16s}
\alias{WhamFrom16s}
\title{Workflow Hub for Automated Metagenomic Exploration (WHAM!)}
\usage{
WhamFrom16s(
  taxaCount = NULL,
  colData = NULL,
  design = NULL,
  phyloseq_object = NULL
)
}
\arguments{
\item{taxaCount}{a data frame or a matrix contains the counts generated from the 16s process. The first column of the count table have to be "Taxa"(case insensitive)}

\item{colData}{metadata table contains experiment design with at least one column. The row number of coldata has to be equal to the columns of count table excluding the first three columns.}

\item{design}{a formula expresses how the counts for each genes depend on the variables.e.g ~ group}

\item{phyloseq_object}{object from package phyloseq}
}
\value{
a WhamInput object contains countData, ColData, design, Taxainfo, FeatureInfo, TaxaTable
}
\description{
R package allowing for exploratory analysis of metagenomics and metatranscriptomic data.  Includes visualization and statistical analysis on the gene family and taxa level. This package is developped based on package ALDEx2.
}
\examples{
# see the vignette for more details
}
