% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wham.core.R
\name{WhamFromBiobakery}
\alias{WhamFromBiobakery}
\title{Workflow Hub for Automated Metagenomic Exploration (WHAM!)}
\usage{
WhamFromBiobakery(countData = NULL, colData = NULL, design = NULL)
}
\arguments{
\item{countData}{a data frame or a matrix contains the counts generated from the biobakery process. The first three columns of the count table have to be "Acc","Taxa",and "Feature" respectively(case insensitive)}

\item{colData}{metadata table contains experiment design with at least one column. The row number of coldata has to be equal to the columns of count table excluding the first three columns.}

\item{design}{a formula expresses how the counts for each genes depend on the variables.e.g ~ group}
}
\value{
a WhamInput object contains countData, ColData, design, Taxainfo, FeatureInfo, TaxaTable
}
\description{
R package allowing for exploratory analysis of metagenomics and metatranscriptomic data.  Includes visualization and statistical analysis on the gene family and taxa level. This package is developped based on package ALDEx2.
}
\examples{
# see the vignette for more details
}
