% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wham.core.R
\name{WhamFromEBI}
\alias{WhamFromEBI}
\title{Workflow Hub for Automated Metagenomic Exploration (WHAM!)}
\usage{
WhamFromEBI(
  DE = c("feature", "taxa"),
  featureCount = NULL,
  taxaCount = NULL,
  colData = NULL,
  design = NULL
)
}
\arguments{
\item{DE}{required. Type of differential expression. Options are "taxa",or "feature".}

\item{featureCount}{a data frame or a matrix contains the feature counts generated from the EBI process. The first two columns of the count table have to be "Acc" ,and "Feature" respectively(case insensitive)}

\item{taxaCount}{a data frame or a matrix contains the taxa counts generated from the EBI process. The first column of the count table have to be "taxa" respectively(case insensitive)}

\item{colData}{metadata table contains experiment design with at least one column. The row number of coldata has to be equal to the columns of count table excluding the first three columns.}

\item{design}{a formula expresses how the counts for each genes depend on the variables.e.g ~ group}
}
\value{
a WhamInput object contains taxaCount/featureCount, ColData, design, Taxainfo, FeatureInfo, TaxaTable
}
\description{
Workflow Hub for Automated Metagenomic Exploration (WHAM!)
}
\examples{
# see the vignette for more details
}
