% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wham.core.R
\name{WhamInputfrom16s}
\alias{WhamInputfrom16s}
\title{Workflow Hub for Automated Metagenomic Exploration (WHAM!)}
\usage{
WhamInputfrom16s(se = NULL, design = NULL, TaxaInfo = NULL, TaxaTable = NULL)
}
\arguments{
\item{se}{a object whose class is the RangedSummarizedExperiment generated from package SummarizedExperiment. Columns of variables indicate sample information in ColData.}

\item{design}{a formula expresses how the counts for each genes depend on the variables.e.g ~ group}

\item{TaxaInfo}{a vector indicates the taxanomic information. The length of the vector is equal to the row number of count table.}

\item{TaxaTable}{a data frame whose columns are kindom, phylum, class, order, family, genus and species indicate the taxonomic information. The row number of TaxaTable is equal to the row number of count table}
}
\value{
a WhamInput object contains all input information
}
\description{
R package allowing for exploratory analysis of metagenomics and metatranscriptomic data.  Includes visualization and statistical analysis on the gene family and taxa level. This package is developped based on package ALDEx2.
}
\examples{
# see the vignette for more details
}
