% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wham.core.R
\name{WhamInputfromEBI}
\alias{WhamInputfromEBI}
\title{Workflow Hub for Automated Metagenomic Exploration (WHAM!)}
\usage{
WhamInputfromEBI(
  se = NULL,
  design = NULL,
  TaxaInfo = NULL,
  FeatureInfo = NULL,
  TaxaTable = NULL,
  featureCount = NULL,
  taxaCount = NULL
)
}
\arguments{
\item{se}{a object whose class is the RangedSummarizedExperiment generated from package SummarizedExperiment. Columns of variables indicate sample information in ColData.}

\item{design}{a formula expresses how the counts for each genes depend on the variables.e.g ~ group}

\item{TaxaInfo}{a vector indicates the taxanomic information. The length of the vector is equal to the row number of count table.}

\item{FeatureInfo}{a vector indicates the genomic information. The length of the vector is equal to the row number of count table.}

\item{TaxaTable}{a data frame whose columns are kindom, phylum, class, order, family, genus and species indicate the taxanomic information. The row number of TaxaTable is equal to the row number of count table}

\item{featureCount}{a data frame or a matrix contains the feature counts generated from the EBI process. The first two columns of the count table have to be "Acc" ,and "Feature" respectively(case insensitive)}

\item{taxaCount}{a data frame or a matrix contains the taxa counts generated from the EBI process. The first column of the count table have to be "taxa" respectively(case insensitive)}
}
\value{
a WhamInput object contains all input information
}
\description{
R package allowing for exploratory analysis of metagenomics and metatranscriptomic data.  Includes visualization and statistical analysis on the gene family and taxa level. This package is developped based on package ALDEx2.
}
\examples{
# see the vignette for more details
}
