% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convergence.R
\name{check_convergence}
\alias{check_convergence}
\title{Check convergence of WHAM model}
\usage{
check_convergence(mod, ret = FALSE, f = "")
}
\arguments{
\item{mod}{output from \code{\link{fit_wham}}}

\item{ret}{T/F, return list? Default = FALSE, just prints to console}
}
\value{
a list with at least the first three of these components:
  \describe{
    \item{\code{$convergence}}{From \code{\link[stats:nlminb]{stats::nlminb}}, "0 indicates successful convergence for nlminb"}
    \item{\code{$maxgr}}{Max absolute gradient value, from `max(abs(mod$gr(mod$opt$par)))`}
    \item{\code{$maxgr_par}}{Name of parameter with max gradient}
    \item{\code{$is_sdrep}}{If \code{\link[TMB:sdreport]{TMB::sdreport}} was performed
    for this model, this indicates whether it performed without error}
    \item{\code{$na_sdrep}}{If \code{\link[TMB:sdreport]{TMB::sdreport}} was performed
    without error for this model, this indicates which (if any) components of the
    diagonal of the inverted hessian were returned as NA}
  }
}
\description{
Access quick convergence checks from `TMB` and `nlminb`.
}
\examples{
\dontrun{
data("input4_SNEMAYT") # load SNEMA yellowtail flounder data and parameter settings
mod = fit_wham(input4_SNEMAYT) # using default values
check_convergence(mod)
}
}
\seealso{
\code{\link{fit_wham}}, \code{\link{fit_tmb}}, \code{\link[stats:nlminb]{stats::nlminb}}
}
