% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tmb.R
\name{fit_tmb}
\alias{fit_tmb}
\title{Fit TMB model using nlminb}
\usage{
fit_tmb(
  model,
  n.newton = 3,
  do.sdrep = TRUE,
  do.check = FALSE,
  save.sdrep = FALSE,
  use.optim = FALSE,
  opt.control = NULL
)
}
\arguments{
\item{model}{Output from \code{\link[TMB:MakeADFun]{TMB::MakeADFun}}.}

\item{n.newton}{Integer, number of additional Newton steps after optimization. Default = \code{3}.}

\item{do.sdrep}{T/F, calculate standard deviations of model parameters? See \code{\link[TMB]{TMB::sdreport}}. Default = \code{TRUE}.}

\item{do.check}{T/F, check if model parameters are identifiable? Runs internal \code{check_estimability}, originally provided by https://github.com/kaskr/TMB_contrib_R/TMBhelper. Default = \code{TRUE}.}

\item{save.sdrep}{T/F, save the full \code{\link[TMB]{TMB::sdreport}} object? If \code{FALSE}, only save \code{\link[TMB:summary.sdreport]{summary.sdreport)}} to reduce model object file size. Default = \code{FALSE}.}

\item{use.optim}{T/F, use \code{\link[stats]{stats::optim}} instead of \code{\link[stats]{stats::nlminb}}? Default = \code{FALSE}.}

\item{opt.control}{list of control parameters to pass to optimizer. For nlminb default = list(iter.max = 1000, eval.max = 1000). For optim default = list(maxit=1000).}
}
\value{
\code{model}, appends the following:
  \describe{
    \item{\code{model$opt}}{Output from \code{\link[stats:nlminb]{stats::nlminb}}}
    \item{\code{model$date}}{System date}
    \item{\code{model$dir}}{Current working directory}
    \item{\code{model$rep}}{model$report()}
    \item{\code{model$TMB_version}}{Version of TMB installed}
    \item{\code{model$parList}}{List of parameters, \code{model$env$parList()}}
    \item{\code{model$final_gradient}}{Final gradient, \code{model$gr()}}
    \item{\code{model$sdrep}}{Estimated standard deviations for model parameters, \code{\link[TMB:sdreport]{TMB::sdreport}} or \code{\link[TMB:summary.sdreport]{summary.sdreport)}}}
  }
}
\description{
Runs optimization on the TMB model using \code{\link[stats:nlminb]{stats::nlminb}}.
If specified, takes additional Newton steps and calculates standard deviations.
Internal function called by \code{\link{fit_wham}}.
}
\seealso{
\code{\link{fit_wham}}, \code{\link{retro}}, \code{TMBhelper::check_estimability}
}
