% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_osa_residuals.R
\name{make_osa_residuals}
\alias{make_osa_residuals}
\title{Calculate one-step-ahead residuals}
\usage{
make_osa_residuals(
  model,
  osa.opts = list(method = "oneStepGaussianOffMode", parallel = TRUE),
  sdrep_required = TRUE
)
}
\arguments{
\item{model}{A fit WHAM model, output from \code{\link{fit_wham}}.}
}
\value{
the same fit TMB model with additional elements for osa residuals:
  \describe{
    \item{\code{$OSA.Ecov}}{data.frame returned by \code{\link[TMB]{TMB::oneStepPredict}} for environmental observations, if applicable.}
    \item{\code{$OSA.agregate}}{data.frame returned by \code{\link[TMB]{TMB::oneStepPredict}} for aggregate catch and index 
       observations conditional on any environmental observations, if applicable.}
    \item{\code{$OSA.agecomp}}{data.frame returned by \code{\link[TMB]{TMB::oneStepPredict}} for age composition observations 
       conditional on any aggregate catch or index, or environmental observations, if applicable.}
    \item{\code{$osa}}{One-step-ahead residuals (if \code{do.osa=TRUE})}
  }
}
\description{
Standard residuals are not appropriate for models with random effects. Instead, one-step-ahead (OSA) residuals
can be used for evaluating model goodness-of-fit (\href{https://doi.org/10.1007/s10651-017-0372-4}{Thygeson et al. (2017)},
implemented in \code{\link[TMB:oneStepPredict]{TMB::oneStepPredict}}). OSA residual options
are passed to \code{\link[TMB:oneStepPredict]{TMB::oneStepPredict}} in a list \code{osa.opts}. Current options are method: 
oneStepGaussianOffMode (default), oneStepGaussian, or oneStepGeneric, and parallel: TRUE/FALSE. 
See \code{\link[TMB:oneStepPredict]{TMB::oneStepPredict}} for further details.
It is not recommended to run this function (or \code{\link[TMB:oneStepPredict]{TMB::oneStepPredict}}) with any random effects and
mvtweedie age composition likelihoods due to extensive computational demand. An error will be thrown in such cases. 
See \href{https://doi.org/10.1016/j.fishres.2022.106487}{Trijoulet et al. (2023)} for OSA methods for age composition OSA residuals.
}
\examples{
\dontrun{
data("input4_SNEMAYT") # load SNEMA yellowtail flounder data and parameter settings
mod <- fit_wham(input4_SNEMAYT, do.osa =FALSE, do.retro =FALSE)
mod <- make_osa_residuals(mod) # calculate Mohn's rho
plot_wham_output(mod)
}

}
\seealso{
\code{\link{fit_wham}}
}
