% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_wham_plots.R
\name{make_wham_plots}
\alias{make_wham_plots}
\title{Make WHAM plots}
\usage{
make_wham_plots(mod, dir = getwd(), out.type = "html")
}
\arguments{
\item{mod}{output from \code{\link{fit_wham}}}

\item{dir}{character, directory to save plots to (default = \code{getwd()})}

\item{out.type}{character, either 'html', 'pdf', or 'png' (default = 'html')}
}
\description{
Generates many output plots for a fit WHAM model.
}
\details{
\code{out.type = 'html'} is the default, makes an html file viewable in any
browser (tabs: 'input data', 'diagnostics', 'results', 'SPR / MSY', 'retro',
and 'misc'). \code{out.type = 'pdf'} makes one pdf file of all plots.
\code{out.type = 'png'} creates a subdirectory 'plots_png' in \code{dir} and
saves .png files within.

Plot functions are located in \code{wham_plots_tables.R}
}
\examples{
\dontrun{
data("SNEMA_ytl") # load SNEMA yellowtail flounder data and parameter settings
mod = fit_wham(input)
make_wham_plots(mod)
}
}
\seealso{
\code{\link{fit_wham}}, \code{wham_plots_tables}
}
