% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wham_output.R
\name{plot_wham_output}
\alias{plot_wham_output}
\title{Plot WHAM output}
\usage{
plot_wham_output(mod, dir.main = getwd(), out.type = "html", res = 72)
}
\arguments{
\item{mod}{output from \code{\link{fit_wham}}}

\item{dir.main}{character, directory to save plots to (default = \code{getwd()})}

\item{out.type}{character, either \code{'html'}, \code{'pdf'}, or \code{'png'} (default = \code{'html'})}

\item{res}{resolution to save .png files (dpi)}
}
\description{
Generates many output plots for a fit WHAM model.
}
\details{
\code{out.type = 'pdf'} makes one pdf file of all plots. \code{out.type = 'png'}
creates a subdirectory `plots_png`` in \code{dir.main} and saves .png files within.
\code{out.type = 'html'} (default) makes an html file for viewing these .png files in a browser
(tabs: 'input data', 'diagnostics', 'results', 'ref_points', 'retro', and 'misc').

Plot functions are located in \code{wham_plots_tables.R}
}
\examples{
\dontrun{
data("input4_SNEMAYT") # load fit wham model
mod <- fit_wham(input4_SNEMAYT)
plot_wham_output(mod)
}
}
\seealso{
\code{\link{fit_wham}}, \code{\link{wham_html}}, \code{wham_plots_tables}
}
