% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_projection.R
\name{prepare_projection}
\alias{prepare_projection}
\title{Prepare input data and parameters to project an already fit WHAM model}
\usage{
prepare_projection(model, proj.opts)
}
\arguments{
\item{model}{a previously fit wham model}

\item{proj.opts}{a named list with the following components:
\itemize{
  \item \code{$n.yrs} (integer), number of years to project/forecast. Default = \code{3}.
  \item \code{$use.last.F} (T/F), use terminal year F for projections. Default = \code{TRUE}.
  \item \code{$use.FXSPR} (T/F), calculate F at X% SPR for projections.
  \item \code{$proj.F} (vector), user-specified fishing mortality for projections. Length must equal \code{n.yrs}.
  \item \code{$proj.catch} (vector), user-specified aggregate catch for projections. Length must equal \code{n.yrs}.
  \item \code{$avg.yrs} (vector), specify which years to average over for calculating reference points. Default = last 5 model years, \code{tail(model$years, 5)}.
  \item \code{$cont.Ecov} (T/F), continue Ecov process (e.g. random walk or AR1) for projections. Default = \code{TRUE}.
  \item \code{$use.last.Ecov} (T/F), use terminal year Ecov for projections.
  \item \code{$avg.Ecov.yrs} (vector), specify which years to average over the environmental covariate(s) for projections.
  \item \code{$proj.Ecov} (vector), user-specified environmental covariate(s) for projections. Length must equal \code{n.yrs}.
  \item \code{$cont.Mre} (T/F), continue M random effects (i.e. AR1_y or 2D AR1) for projections. Default = \code{TRUE}. If \code{FALSE}, M will be averaged over \code{$avg.yrs} (which defaults to last 5 model years).
}}
}
\value{
same as \code{\link{prepare_wham_input}}, a list ready for \code{\link{fit_wham}}:
  \describe{
    \item{\code{data}}{Named list of data, passed to \code{\link[TMB:MakeADFun]{TMB::MakeADFun}}}
    \item{\code{par}}{Named list of parameters, passed to \code{\link[TMB:MakeADFun]{TMB::MakeADFun}}}
    \item{\code{map}}{not sure what this does, passed to \code{\link[TMB:MakeADFun]{TMB::MakeADFun}}}
    \item{\code{random}}{Character vector of parameters to treat as random effects, passed to \code{\link[TMB:MakeADFun]{TMB::MakeADFun}}}
    \item{\code{years}}{Numeric vector of years to fit WHAM model (specified in ASAP3 .dat file)}
    \item{\code{ages.lab}}{Character vector of age labels, ending with plus-group (specified in ASAP3 .dat file)}
    \item{\code{model_name}}{Character, name of stock/model (specified in call to \code{prepare_wham_input})}
  }
}
\description{
\code{prepare_projection} is an internal function called by \code{\link{project_wham}},
which in turn is called by \code{\link{fit_wham}} if \code{do.proj = TRUE}.
}
\seealso{
\code{\link{prepare_wham_input}}, \code{\link{project_wham}}
}
