% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_wham.R
\name{project_wham}
\alias{project_wham}
\title{Project a fit WHAM model}
\usage{
project_wham(model, proj.opts = list(n.yrs = 3, use.last.F = TRUE,
  use.avg.F = FALSE, use.FXSPR = FALSE, proj.F = NULL, proj.catch = NULL,
  avg.yrs = NULL, cont.Ecov = TRUE, use.last.Ecov = FALSE, avg.Ecov.yrs =
  NULL, proj.Ecov = NULL), n.newton = 3, do.sdrep = TRUE)
}
\arguments{
\item{model}{a previously fit wham model}

\item{proj.opts}{a named list with the following components:
\itemize{
  \item \code{$n.yrs} (integer), number of years to project/forecast. Default = \code{3}.
  \item \code{$use.last.F} (T/F), use terminal year F for projections. Default = \code{TRUE}.
  \item \code{$use.FXSPR} (T/F), calculate F at X% SPR for projections.
  \item \code{$proj.F} (vector), user-specified fishing mortality for projections. Length must equal \code{n.yrs}.
  \item \code{$proj.catch} (vector), user-specified aggregate catch for projections. Length must equal \code{n.yrs}.
  \item \code{$avg.yrs} (vector), specify which years to average over for calculating reference points. Default = last 5 model years, \code{tail(model$years, 5)}.
  \item \code{$cont.Ecov} (T/F), continue Ecov process (e.g. random walk or AR1) for projections. Default = \code{TRUE}.
  \item \code{$use.last.Ecov} (T/F), use terminal year Ecov for projections.
  \item \code{$avg.Ecov.yrs} (vector), specify which years to average over the environmental covariate(s) for projections.
  \item \code{$proj.Ecov} (matrix), user-specified environmental covariate(s) for projections. \code{n.yrs} rows.
}}

\item{n.newton}{integer, number of additional Newton steps after optimization. Passed to \code{\link{fit_tmb}}. Default = \code{0} for projections.}

\item{do.sdrep}{T/F, calculate standard deviations of model parameters? See \code{\link[TMB]{sdreport}}. Default = \code{TRUE}.}
}
\value{
a projected WHAM model with additional output if specified:
  \describe{
    \item{\code{$rep}}{List of derived quantity estimates (see examples)}
    \item{\code{$sdrep}}{Parameter estimates (and standard errors if \code{do.sdrep=TRUE})}
    \item{\code{$peels}}{Retrospective analysis (if \code{do.retro=TRUE})}
    \item{\code{$osa}}{One-step-ahead residuals (if \code{do.osa=TRUE})}
  }
}
\description{
Provides projections/forecasts for an existing (already fit) WHAM model.
}
\details{
WHAM implements five options for handling fishing mortality in the projections.
Exactly one of these must be specified in \code{proj.opts}:
  \itemize{
    \item Use last year F (default). Set \code{proj.opts$use.last.F = TRUE}. WHAM will use F in the terminal model year for projections.
    \item Use average F. Set \code{proj.opts$use.avg.F = TRUE}. WHAM will use F averaged over \code{proj.opts$avg.yrs} for projections (as is done for M-, maturity-, and weight-at-age).
    \item Use F at X% SPR. Set \code{proj.opts$use.FXSPR = TRUE}. WHAM will calculate F at X% SPR.
    \item Specify F. Provide \code{proj.opts$proj.F}, an F vector with length = \code{n.yrs}.
    \item Specify catch. Provide \code{proj.opts$proj.catch}, a vector of aggregate catch with length = \code{n.yrs}. WHAM will calculate F to get specified catch.
  }

\code{proj.opts$avg.yrs} controls which years will be averaged over in the projections.
The following quantities are averaged:
  \itemize{
    \item Maturity-at-age
    \item Weight-at-age
    \item Natural mortality-at-age
    \item Fishing mortality-at-age (if \code{proj.opts$use.avgF = TRUE})
  }

WHAM implements four options for handling the environmental covariate(s) in the projections.
Exactly one of these must be specified in \code{proj.opts} if \code{Ecov} is in the model:
  \describe{
    \item{(Default) Continue Ecov process model (e.g. random walk, AR1)}{Set \code{$cont.Ecov = TRUE}. WHAM will estimate the Ecov process in projection years (i.e. continue the random walk / AR1 process).}
    \item{Use last year Ecov(s)}{Set \code{$use.last.Ecov = TRUE}. WHAM will use Ecov value from the terminal year (of population model) for projections.}
    \item{Use average Ecov(s)}{Provide \code{$avg.yrs.Ecov}, a vector specifying which years to average over the environmental covariate(s) for projections.}
    \item{Specify \code{Ecov}}{Provide \code{$proj.Ecov}, a matrix of user-specified environmental covariate(s) to use for projections. Dimensions must be # projection years (\code{proj.opts$n.yrs}) x # Ecovs (\code{ncols(Ecov$mean)}).}
  }
If the original model fit the Ecov in years beyond the population model, WHAM will use the already-fit
Ecov values for the projections. If the Ecov model extended at least \code{proj.opts$n.yrs} years
beyond the population model, then none of the above need be specified.
}
\examples{
\dontrun{
data("SNEMA_ytl") # load SNEMA yellowtail flounder data and parameter settings
mod <- fit_wham(input) # using default values (do.proj=T)

mod2 <- fit_wham(input, do.retro=F, do.osa=F, do.proj=F) # fit model without projections, retro analysis, or OSA residuals
mod_proj <- project_wham(mod2) # add projections to previously fit model, using default values: use.lastF = TRUE, n.yrs = 3, avg.yrs = last 5 years

names(mod_proj$rep) # list of derived quantities
tail(mod_proj$rep$SSB, 3) # get 3-year projected SSB estimates (weight, not numbers)
x = summary(mod_proj$sdrep)
unique(rownames(x))) # list of estimated parameters and derived quanitites with SE
x = x[rownames(x) == "log_SSB",] # SSB estimates with SE
ssb.mat = exp(cbind(x, x[,1] + qnorm(0.975)*cbind(-x[,2],x[,2])))/1000 # calculate 95\% CI
colnames(ssb.mat) <- c("SSB","SSB_se","SSB_lower","SSB_upper")
tail(ssb.mat, 3) # 3-year projected SSB estimates with SE and 95\% CI
}
}
\seealso{
\code{\link{fit_wham}}, \code{\link{fit_tmb}}
}
