% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_asap3_dat.R
\name{read_asap3_dat}
\alias{read_asap3_dat}
\title{Read an ASAP3 .dat file into R}
\usage{
read_asap3_dat(filename)
}
\arguments{
\item{filename}{character vector, names of ASAP3 .dat files. The file either needs to be
in the current working directory, or \code{filename} can include the path. If multipile files,
a multi-stock model will be assumed.}
}
\value{
a named list with the following components:
  \describe{
    \item{\code{dat}}{Named list of input data and parameters}
    \item{\code{comments}}{Comments at top of ASAP3 .dat file (indicated by "#")}
  }
}
\description{
WHAM is built on ASAP ([Legault and Restrepo (1999)](http://sedarweb.org/docs/wsupp/S12RD06\%20ASAPdoc.pdf)) and this
function provides functionality to use a preexisting ASAP3 input data file. The
output of \code{read_asap3_dat} should then be passed to \code{\link{prepare_wham_input}}.
If you are not familiar with ASAP3 input files, see the ASAP \href{https://github.com/cmlegault/ASAPplots/tree/master/pdf}{documentation}
and \href{https://nmfs-fish-tools.github.io/ASAP/}{code}.
}
\examples{
\dontrun{
asap3 = read_asap3_dat("ASAP_SNEMAYT.dat")
input = prepare_wham_input(asap3)
mod = fit_wham(input)
}

}
\seealso{
\code{\link{prepare_wham_input}}, \code{\link{fit_wham}}, \href{https://github.com/cmlegault/ASAPplots/tree/master/pdf}{ASAP documentation}
}
