% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro.R
\name{retro}
\alias{retro}
\title{Run retrospective analysis}
\usage{
retro(model, n.peels = 7, ran = "log_NAA", do.sdrep = FALSE, n.newton = 0)
}
\arguments{
\item{model}{Optimized TMB model, output from \code{\link{fit_tmb}}.}

\item{n.peels}{Integer, number of peels to use in retrospective analysis. Default = \code{7}.}

\item{ran}{Character, specifies which parameters to treat as random effects. Default = \code{"log_NAA"}.}

\item{do.sdrep}{T/F, calculate standard deviations of model parameters for each peel? Default = \code{FALSE}.}

\item{n.newton}{integer, number of additional Newton steps after optimization for each peel. Default = \code{0}.}
}
\value{
\code{peels}, a list of length \code{n.peels}, where entry \emph{i} is a model
fit by peeling off \emph{i} years of data.
}
\description{
Internal function called by \code{\link{fit_wham}}. Calls \code{\link{fit_peel}}
to fit the model peeling off \code{1, 2, ..., n.peels} years of data.
}
\seealso{
\code{\link{fit_wham}}, \code{\link{fit_peel}}
}
