% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_selectivity.R
\name{set_selectivity}
\alias{set_selectivity}
\title{Specify model and parameter configuration for selectivity}
\usage{
set_selectivity(input, selectivity)
}
\arguments{
\item{input}{list containing data, parameters, map, and random elements (output from \code{\link{prepare_wham_input}})}

\item{selectivity}{(optional) list specifying options for selectivity blocks, models, initial parameter values, parameter fixing/mapping, and random effects (see details)

\code{set_selectivity} specifies options for selectivity and allows you to overwrite existing options 
in the \code{input} list or as specified in the ASAP data file. If \code{selectivity = NULL}, selectivity options from 
\code{input} are used. 

\code{\link{prepare_wham_input}(..., selectivity=selectivity)} calls \code{set_selectivity(..., selectivity=selectivity)}.
If you already have created \code{input} with \code{prepare_wham_input}, you can also use \code{set_selectivity(input, selectivity=selectivity)}
to modify the selectivity specification.

\code{selectivity} is a list with the following entries:
  \describe{
    \item{$model}{Selectivity model for each block. Vector with length = number of selectivity blocks. Each entry must be one of: "age-specific", "logistic", "double-logistic", or "decreasing-logistic".}
    \item{$re}{Time-varying (random effects) for each block. Vector with length = number of selectivity blocks.
                 If \code{NULL}, selectivity parameters in all blocks are constant over time and uncorrelated.
                 Each entry of \code{selectivity$re} must be one of the following options, where the selectivity parameters are:
                 \describe{
                   \item{"none"}{(default) are constant and uncorrelated}
                   \item{"iid"}{vary by year and age/par, but uncorrelated}
                   \item{"ar1"}{correlated by age/par (AR1), but not year}
                   \item{"ar1_y"}{correlated by year (AR1), but not age/par}
                   \item{"2dar1"}{correlated by year and age/par (2D AR1)}
                 }
                }
    \item{$initial_pars}{Initial parameter values for each block. List of length = number of selectivity blocks. Each entry must be
               a vector of length # parameters in the block, i.e. \code{c(2,0.2)} for logistic (a50 and 1/slope) or \code{c(0.5,0.5,0.5,1,1,0.5)} for
               age-specific parameters when there are 6 ages. Default is to set at middle of parameter range. This is 0.5 for age-specific and n.ages/2 
               or logistic, double-logistic, and decreasing-logistic.}
    \item{$fix_pars}{Alternative to \code{$map_pars} for specifying which selectivity parameters (only fixed effects) to fix at initial values. 
               List of length = number of selectivity blocks. E.g. model with 3 age-specific blocks and 6 ages, 
               \code{list(4:5, 4, 2:4))} will fix ages 4 and 5 in block 1, age 4 in block 2, and ages 2, 3, and 4 in block 3.
               Use NULL to not fix any parameters for a block, e.g. list(NULL, 4, 2) does not fix any pars in block 1.}
    \item{$par_min}{The lower bound for selectivity parameters and is used to populate \code{data$selpars_lower}.
               List of length = number of selectivity blocks, where each item is a 
               vector of length = number of selectivity parameters (age-specific: n.ages, logistic: 2, 
               double-logistic: 4).}
    \item{$par_max}{The upper bound for selectivity parameters and is used to populate \code{data$selpars_upper}.
               List of length = number of selectivity blocks, where each item is a 
               vector of length = number of selectivity parameters (age-specific: n.ages, logistic: 2, 
               double-logistic: 4).}
    \item{$map_pars}{Alternative to \code{$fix_pars} for specifying how to fix selectivity parameters (only fixed effects), corresponds 
               to \code{map$logit_selpars}. List of length = number of selectivity blocks, where each item is a 
               vector of length = number of selectivity parameters (age-specific: n.ages, logistic: 2, 
               double-logistic: 4). Use \code{NA} to fix a parameter and integers to estimate. Use the same integer
               for multiple ages or fleets/indices to estimate a shared parameter. E.g. for a model with 3 age-specific 
               blocks (1 fleet, 2 indices) and 6 ages, \code{$map_pars = list(c(1,2,3,NA,NA,4), c(5,6,7,NA,8,8), c(1,2,3,NA,NA,4))}
               will estimate ages 1-3 and 6 in block 1 (fleet), ages 1-3 and 4-5 (shared) in block 2 (index 1), and then set the
               index 2 (block 3) selectivity equal to the fleet.}
    \item{$sigma_vals}{Initial standard deviation values to use for the random effect deviations. Must be a vector with length = number of blocks. 
               Use natural (not log) scale, must be positive. \code{par$sel_repars[,1]} will be estimated on log-scale. Not used if \code{re = 'none'} for all blocks.}
    \item{$map_sigma}{Specify which SD parameters to fix for the random effect deviations. Must be a vector with length = number of blocks. 
               Use \code{NA} to fix a parameter and integers to estimate. Use the same integer for multiple blocks to estimate a shared SD parameter.
               Not used if \code{re = 'none'} for all blocks.}
    \item{$cor_vals}{Initial correlation values to use for the random effect deviations. Must be a n_selblocks x 2 integer matrix. Columns correspond to correlation by age and year,
               respectively. If \code{re = 'ar1'} or \code{re = 'ar1_y'} only the corresponding values are used.
               Values must be between -1 and 1, but parameters are estimated on a logit transformed scale internally. 
               Not used if \code{re = 'none'} or \code{re = 'iid'} for all blocks.}
    \item{$map_cor}{Specify which correlation parameters to fix for the random effect deviations. Must be a n_selblocks x 2 matrix. Columns correspond to correlation by age
               and year,respectively. Parameters can be shared by setting corresponding values of \code{map_cor} to the same integer. Use \code{NA} to fix a parameter.
               If \code{re = 'ar1'} or \code{re = 'ar1_y'}, only the column for the corresponding correlation are used.
               Not used if \code{re = 'none'} or \code{re = 'iid'} for all blocks.}
    \item{$n_selblocks}{How many selectivity blocks. Optional. If unspecified and no asap3 object, then this is set to the number 
               of fleets + indices. If specified, ensure other components of \code{selectivity} are consistent.}
  }}
}
\value{
a named list with same elements as the input provided with selectivity options modified.
}
\description{
Specify model and parameter configuration for selectivity
}
\examples{
\dontrun{
wham.dir <- find.package("wham")
path_to_examples <- system.file("extdata", package="wham")
asap3 <- read_asap3_dat(file.path(path_to_examples,"ex1_SNEMAYT.dat"))
input <- prepare_wham_input(asap3, NAA_re = list(sigma = "rec"))
sel <- list(model=rep("logistic",input$data$n_selblocks),
   initial_pars=rep(list(c(3,3)),input$data$n_selblocks),
   fix_pars=rep(list(NULL),input$data$n_selblocks))
input <- set_selectivity(input, selectivity = sel) #logistic selectivity for all selectivity blocks
}

}
\seealso{
\code{\link{prepare_wham_input}}
}
