% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_om_fn.R
\name{update_om_fn}
\alias{update_om_fn}
\title{Update the operating model and generate data}
\usage{
update_om_fn(om, om2, interval.info, seed = 123)
}
\arguments{
\item{om}{Operating model with years including burn-in + all feedback years}

\item{om2}{Operating model with years including burn-in years + n feedback years (n = the number of assessments completed x assessment interval)}

\item{interval.info}{Catch advice for a number of years projected from the estimation model
\itemize{
  \item \code{"$years"} projection years 
  \item \code{"$catch"} matrix (n_region x n_years) projected catch 
  \item \code{"=NULL"} generate data in the OM
  }}

\item{seed}{Seed used to generate data}
}
\value{
an operating model with simulated data and updated F time series
}
\description{
Function to update F in the operating model (see \code{\link{update_om_F}}) and generate data given the updated F.
}
\seealso{
\code{\link{update_om_F}}
}
